brainstorm-workflow-automatisierung/poc/prefect/requirements.txt
prefect>=2.14.0
requests>=2.31.0
python-dotenv>=1.0.0

